refactor: move api calls to Block page BlockAPI.ts
